Task 2 – Backend Extensions & Data Lift - CODEX HIGH (automagik-forge 3a010ecd)
# Task 2 – Backend Extensions & Data Lift

<task_breakdown>
1. [Discovery] Inventory forge-only backend logic
   - Use diff artifacts and Task 2 prompt to list Omni, branch template, config v7, Genie, and MCP touchpoints.
   - Map which modules migrate into `forge-extensions/*` versus adapters inside `forge-app`.
   - Inspect existing migrations and data models to plan auxiliary table coverage.

2. [Implementation] Extract features into extension crates
   - Relocate forge-only services into dedicated crates (`forge-extensions/omni`, `forge-extensions/branch-templates`, `forge-extensions/config`, `forge-extensions/genie`).
   - Introduce auxiliary tables, views, and data migrations under `forge-app/migrations` with idempotent guards.
   - Replace direct modifications in upstream crates with composition hooks inside `forge-app`.

3. [Verification] Regression-ready backend
   - Seed local snapshot via `./scripts/collect-forge-snapshot.sh` and load data into new tables.
   - Run targeted tests (`cargo test -p forge-extensions-*`, `sqlx migrate run --dry-run`) and backend smoke calls.
   - Update regression docs if API payloads change; otherwise ensure diff against baseline passes.
</task_breakdown>

[SUCCESS CRITERIA]
✅ All forge backend functionality lives in extension crates with clean APIs.
✅ Auxiliary tables/views replicate existing data without loss; migrations are reversible.
✅ `forge-app` compiles and routes requests through composition layer.
✅ Regression harness backend section passes using snapshot data.

[NEVER DO]
❌ Modify code under `upstream/`—extend via adapters only.
❌ Drop functionality (Omni, Genie, branch templates, config v7) unless replaced equivalently.
❌ Ship non-idempotent or destructive migrations.
❌ Commit snapshot data or perform git pushes/PRs.

[CONTEXT]
@genie/wishes/restructure-upstream-library-wish.md
@prompt-task2.md
@docs/regression/baseline/README.md
@docs/regression/logs/cargo-test-workspace.log
@docs/regression/logs/pnpm-run-build-npx.log
@scripts/collect-forge-snapshot.sh
@scripts/run-forge-regression.sh

[TODO]
- Revisit Task 1 outputs to confirm scaffolding is ready for extensions.
- Enumerate every forge-only backend module and design destination crates/interfaces.
- Implement migration files and data loaders that leverage the hydrated snapshot.
- Wire `forge-app` to compose upstream services with the new extensions.
- Execute validation commands (targeted cargo tests, migration dry run, regression harness backend checks) and capture results.
- Document findings, follow-up risks, and next-step guidance without performing commits or PRs.

## Execution Plan (Codex High)

**Map forge-only backend logic across Omni, branch templates, config v7, Genie/MCP**, including data flows and current dependencies; capture findings in notes for reference.
**Design extension crate APIs plus forge-app composition**: align service boundaries, config structures, and identify required adapters; sketch auxiliary table schema and migration flow.

### Phase Breakdown

**Inventory & Notes**
Inspect `crates/services/src/services/{omni,config,genie}` and `crates/db/src/models/{task,task_attempt}` plus MCP touchpoints; record forge-only behaviours, TS bindings, and DB usage.
Confirm Genie scope for this task (document blockers if deferred).

**Extension Extraction**
For each domain (Omni → `forge-extensions/omni`, Branch Templates → `forge-extensions/branch-templates`, Config v7 → `forge-extensions/config`, Genie → `forge-extensions/genie`), move logic, tests, and types.
Update crate manifests and shared TS generation paths; expose clean public APIs; rewire `forge-app` services to consume them.

**Database Migrations**
Author `forge-app/migrations/001_auxiliary_tables.sql` (tables/views with `IF NOT EXISTS` guards) and `002_migrate_data.sql` (copy + null legacy columns, rollback comments).
Plan data loaders leveraging snapshot tooling.

**Composition & Routes**
Implement `ForgeServices` container and `ForgeTaskService` inside `forge-app/src/services`; integrate upstream services via adapters.
Extend router for `/health` and `/api/forge/{omni,branch-templates}` stub endpoints backed by new services.

**Legacy Cleanup**
Remove forge-only code from original crates, leaving adapters/hooks to the extension layer; ensure no residual Omni/branch-template logic in upstream directories.
Adjust MCP/Genie integration to use extensions or document deferral.

**Verification & Docs**
Run `cargo check --workspace`, targeted tests, `sqlx migrate run --dry-run`, and note results (sandbox limits if blocked).
Update prep/regression docs with migration notes, verification transcripts, and outstanding risks.

### Flow (ASCII)
Inventory → [Design APIs & Schema] → {Extract Extensions} → {Create Migrations} → {Wire forge-app routes/services} → {Cleanup Legacy Paths} → {Verify & Document}
