+++ crates/services/Cargo.toml
@@ -61,3 +61,5 @@ once_cell = "1.20"
 sha2 = "0.10"
 fst = "0.4"
 moka = { version = "0.12", features = ["future"] }
+forge-extensions-omni = { path = "../../forge-extensions/omni" }
+forge-extensions-config = { path = "../../forge-extensions/config" }
