+++ crates/services/src/services/config/mod.rs
@@ -1,46 +1,16 @@
-use std::path::PathBuf;
+pub use forge_extensions_config::versions;
 
-use thiserror::Error;
-
-mod versions;
-
-#[derive(Debug, Error)]
-pub enum ConfigError {
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-    #[error(transparent)]
-    Json(#[from] serde_json::Error),
-    #[error("Validation error: {0}")]
-    ValidationError(String),
-}
-
-pub type Config = versions::v7::Config;
-pub type NotificationConfig = versions::v7::NotificationConfig;
-pub type EditorConfig = versions::v7::EditorConfig;
-pub type ThemeMode = versions::v7::ThemeMode;
-pub type SoundFile = versions::v7::SoundFile;
-pub type EditorType = versions::v7::EditorType;
-pub type GitHubConfig = versions::v7::GitHubConfig;
-pub type OmniConfig = versions::v7::OmniConfig;
-pub type RecipientType = versions::v7::RecipientType;
-
-/// Will always return config, trying old schemas or eventually returning default
-pub async fn load_config_from_file(config_path: &PathBuf) -> Config {
-    match std::fs::read_to_string(config_path) {
-        Ok(raw_config) => Config::from(raw_config),
-        Err(_) => {
-            tracing::info!("No config file found, creating one");
-            Config::default()
-        }
-    }
-}
-
-/// Saves the config to the given path
-pub async fn save_config_to_file(
-    config: &Config,
-    config_path: &PathBuf,
-) -> Result<(), ConfigError> {
-    let raw_config = serde_json::to_string_pretty(config)?;
-    std::fs::write(config_path, raw_config)?;
-    Ok(())
-}
+pub use forge_extensions_config::{
+    load_config_from_file,
+    save_config_to_file,
+    ConfigError,
+    EditorConfig,
+    EditorType,
+    ForgeConfig as Config,
+    GitHubConfig,
+    NotificationConfig,
+    OmniConfig,
+    RecipientType,
+    SoundFile,
+    ThemeMode,
+};
