+++ forge-extensions/branch-templates/Cargo.toml
@@ -6,4 +6,5 @@ publish = false
 
 [dependencies]
 serde = { workspace = true }
-anyhow = { workspace = true }
+uuid = { version = "1.0", features = ["v4"] }
+utils = { path = "../../crates/utils" }
