+++ forge-extensions/config/Cargo.toml
@@ -1,10 +1,20 @@
 [package]
 name = "forge-extensions-config"
 version = "0.0.0"
-edition = "2021"
+edition = "2024"
 publish = false
 
 [dependencies]
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
+anyhow = { workspace = true }
+ts-rs = { workspace = true }
+tracing = { workspace = true }
+executors = { path = "../../crates/executors" }
+forge-extensions-omni = { path = "../omni" }
+strum = "0.27.2"
+strum_macros = "0.27.2"
+utils = { path = "../../crates/utils" }
+rust-embed = "8.2"
+tokio = { workspace = true }
