+++ forge-extensions/config/src/versions/v5.rs
@@ -1,19 +1,15 @@
-use std::str::FromStr;
-
 use anyhow::Error;
-use executors::{executors::BaseCodingAgent, profile::ExecutorProfileId};
 use serde::{Deserialize, Serialize};
 use ts_rs::TS;
-use utils;
-pub use v5::{EditorConfig, EditorType, GitHubConfig, NotificationConfig, SoundFile, ThemeMode};
+pub use v4::{EditorConfig, EditorType, GitHubConfig, NotificationConfig, SoundFile, ThemeMode};
 
-use crate::services::config::versions::v5;
+use super::v4::{self, ProfileVariantLabel};
 
 #[derive(Clone, Debug, Serialize, Deserialize, TS)]
 pub struct Config {
     pub config_version: String,
     pub theme: ThemeMode,
-    pub executor_profile: ExecutorProfileId,
+    pub profile: ProfileVariantLabel,
     pub disclaimer_acknowledged: bool,
     pub onboarding_acknowledged: bool,
     pub github_login_acknowledged: bool,
@@ -29,7 +25,7 @@ pub struct Config {
 
 impl Config {
     pub fn from_previous_version(raw_config: &str) -> Result<Self, Error> {
-        let old_config = match serde_json::from_str::<v5::Config>(raw_config) {
+        let old_config = match serde_json::from_str::<v4::Config>(raw_config) {
             Ok(cfg) => cfg,
             Err(e) => {
                 tracing::error!("❌ Failed to parse config: {}", e);
@@ -38,36 +34,10 @@ impl Config {
             }
         };
 
-        // Backup custom profiles.json if it exists (v6 migration may break compatibility)
-        let profiles_path = utils::assets::profiles_path();
-        if profiles_path.exists() {
-            let backup_name = format!(
-                "profiles_v5_backup_{}.json",
-                std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .unwrap()
-                    .as_secs()
-            );
-            let backup_path = profiles_path.parent().unwrap().join(backup_name);
-
-            if let Err(e) = std::fs::rename(&profiles_path, &backup_path) {
-                tracing::warn!("Failed to backup profiles.json: {}", e);
-            } else {
-                tracing::info!("Custom profiles.json backed up to {:?}", backup_path);
-                tracing::info!("Please review your custom profiles after migration to v6");
-            }
-        }
-
-        // Validate and convert ProfileVariantLabel
-        let old_coding_agent = old_config.profile.profile.to_uppercase();
-        let base_coding_agent =
-            BaseCodingAgent::from_str(&old_coding_agent).unwrap_or(BaseCodingAgent::ClaudeCode);
-        let executor_profile = ExecutorProfileId::new(base_coding_agent);
-
         Ok(Self {
-            config_version: "v6".to_string(),
+            config_version: "v5".to_string(),
             theme: old_config.theme,
-            executor_profile,
+            profile: old_config.profile,
             disclaimer_acknowledged: old_config.disclaimer_acknowledged,
             onboarding_acknowledged: old_config.onboarding_acknowledged,
             github_login_acknowledged: old_config.github_login_acknowledged,
@@ -77,8 +47,8 @@ impl Config {
             github: old_config.github,
             analytics_enabled: old_config.analytics_enabled,
             workspace_dir: old_config.workspace_dir,
-            last_app_version: old_config.last_app_version,
-            show_release_notes: old_config.show_release_notes,
+            last_app_version: None,
+            show_release_notes: false,
         })
     }
 }
@@ -86,14 +56,14 @@ impl Config {
 impl From<String> for Config {
     fn from(raw_config: String) -> Self {
         if let Ok(config) = serde_json::from_str::<Config>(&raw_config)
-            && config.config_version == "v6"
+            && config.config_version == "v5"
         {
             return config;
         }
 
         match Self::from_previous_version(&raw_config) {
             Ok(config) => {
-                tracing::info!("Config upgraded to v6");
+                tracing::info!("Config upgraded to v5");
                 config
             }
             Err(e) => {
@@ -107,9 +77,9 @@ impl From<String> for Config {
 impl Default for Config {
     fn default() -> Self {
         Self {
-            config_version: "v6".to_string(),
+            config_version: "v5".to_string(),
             theme: ThemeMode::System,
-            executor_profile: ExecutorProfileId::new(BaseCodingAgent::ClaudeCode),
+            profile: ProfileVariantLabel::default("claude-code".to_string()),
             disclaimer_acknowledged: false,
             onboarding_acknowledged: false,
             github_login_acknowledged: false,
