+++ forge-extensions/omni/Cargo.toml
@@ -9,3 +9,5 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 tracing = { workspace = true }
 anyhow = { workspace = true }
+ts-rs = { workspace = true }
+reqwest = { version = "0.12", features = ["json"] }
